requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pycodestyle==2.7.0
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
python-dotenv==0.19.2 